requirements.txt
pyodbc==5.2.0
psycopg[binary,pool]==3.2.9
python-dotenv==1.1.1
schedule==1.2.2
psycopg2-binary==2.9.10
 